Handle case where user is not logged in
